Refactor AuthenticationHeaderValueProvider construction
