+++ src/Confluent.SchemaRegistry/CachedSchemaRegistryClient.cs
@@ -23,7 +23,6 @@ using System.Collections.Generic;
 using System.Threading.Tasks;
 using System.Linq;
 using System;
-using System.Net.Http;
 using System.Collections.Concurrent;
 using System.Net;
 using System.Security.Cryptography.X509Certificates;
@@ -308,165 +307,8 @@ namespace Confluent.SchemaRegistry
                     $"Configured value for {SchemaRegistryConfig.PropertyNames.SchemaRegistryLatestCacheTtlSecs} must be an integer.");
             }
             
-            var basicAuthSource = config.FirstOrDefault(prop =>
-                    prop.Key.ToLower() == SchemaRegistryConfig.PropertyNames.SchemaRegistryBasicAuthCredentialsSource)
-                .Value ?? "";
-            var basicAuthInfo = config.FirstOrDefault(prop =>
-                prop.Key.ToLower() == SchemaRegistryConfig.PropertyNames.SchemaRegistryBasicAuthUserInfo).Value ?? "";
-
-            string username = null;
-            string password = null;
-
-            if (basicAuthSource == "USER_INFO" || basicAuthSource == "")
-            {
-                if (basicAuthInfo != "")
-                {
-                    var userPass = basicAuthInfo.Split(new char[] { ':' }, 2);
-                    if (userPass.Length != 2)
-                    {
-                        throw new ArgumentException(
-                            $"Configuration property {SchemaRegistryConfig.PropertyNames.SchemaRegistryBasicAuthUserInfo} must be of the form 'username:password'.");
-                    }
-
-                    username = userPass[0];
-                    password = userPass[1];
-                    if (authenticationHeaderValueProvider != null)
-                    {
-                        throw new ArgumentException(
-                            $"Invalid authentication header value provider configuration: Cannot specify both custom provider and username/password");
-                    }
-                    authenticationHeaderValueProvider = new BasicAuthenticationHeaderValueProvider(username, password);
-                }
-            }
-            else if (basicAuthSource == "SASL_INHERIT")
-            {
-                if (basicAuthInfo != "")
-                {
-                    throw new ArgumentException(
-                        $"{SchemaRegistryConfig.PropertyNames.SchemaRegistryBasicAuthCredentialsSource} set to 'SASL_INHERIT', but {SchemaRegistryConfig.PropertyNames.SchemaRegistryBasicAuthUserInfo} as also specified.");
-                }
-
-                var saslUsername = config.FirstOrDefault(prop => prop.Key == "sasl.username");
-                var saslPassword = config.FirstOrDefault(prop => prop.Key == "sasl.password");
-                if (saslUsername.Value == null)
-                {
-                    throw new ArgumentException(
-                        $"{SchemaRegistryConfig.PropertyNames.SchemaRegistryBasicAuthCredentialsSource} set to 'SASL_INHERIT', but 'sasl.username' property not specified.");
-                }
-
-                if (saslPassword.Value == null)
-                {
-                    throw new ArgumentException(
-                        $"{SchemaRegistryConfig.PropertyNames.SchemaRegistryBasicAuthCredentialsSource} set to 'SASL_INHERIT', but 'sasl.password' property not specified.");
-                }
-
-                username = saslUsername.Value;
-                password = saslPassword.Value;
-                if (authenticationHeaderValueProvider != null)
-                {
-                    throw new ArgumentException(
-                        $"Invalid authentication header value provider configuration: Cannot specify both custom provider and username/password");
-                }
-                authenticationHeaderValueProvider = new BasicAuthenticationHeaderValueProvider(username, password);
-            }
-            else
-            {
-                throw new ArgumentException(
-                    $"Invalid value '{basicAuthSource}' specified for property '{SchemaRegistryConfig.PropertyNames.SchemaRegistryBasicAuthCredentialsSource}'");
-            }
-
-            var bearerAuthSource = config.FirstOrDefault(prop =>
-                prop.Key.ToLower() == SchemaRegistryConfig.PropertyNames.SchemaRegistryBearerAuthCredentialsSource).Value ?? "";
-
-            if (bearerAuthSource != "" && basicAuthSource != "")
-            {
-                throw new ArgumentException(
-                    $"Invalid authentication header value provider configuration: Cannot specify both basic and bearer authentication");
-            }
-
-            string logicalCluster = null;
-            string identityPoolId = null;
-            string bearerToken = null;
-            string clientId = null;
-            string clientSecret = null;
-            string scope = null;
-            string tokenEndpointUrl = null;
-
-            if (bearerAuthSource == "STATIC_TOKEN" || bearerAuthSource == "OAUTHBEARER")
-            {
-                if (authenticationHeaderValueProvider != null)
-                {
-                    throw new ArgumentException(
-                        $"Invalid authentication header value provider configuration: Cannot specify both custom provider and bearer authentication");
-                }
-                logicalCluster = config.FirstOrDefault(prop =>
-                    prop.Key.ToLower() == SchemaRegistryConfig.PropertyNames.SchemaRegistryBearerAuthLogicalCluster).Value;
-
-                identityPoolId = config.FirstOrDefault(prop =>
-                    prop.Key.ToLower() == SchemaRegistryConfig.PropertyNames.SchemaRegistryBearerAuthIdentityPoolId).Value;
-                if (logicalCluster == null || identityPoolId == null)
-                {
-                    throw new ArgumentException(
-                        $"Invalid bearer authentication provider configuration: Logical cluster and identity pool ID must be specified");
-                }
-            }
-
-            switch (bearerAuthSource)
-            {
-                case "STATIC_TOKEN":
-                    bearerToken = config.FirstOrDefault(prop =>
-                        prop.Key.ToLower() == SchemaRegistryConfig.PropertyNames.SchemaRegistryBearerAuthToken).Value;
-
-                    if (bearerToken == null)
-                    {
-                        throw new ArgumentException(
-                            $"Invalid authentication header value provider configuration: Bearer authentication token not specified");
-                    }
-                    authenticationHeaderValueProvider = new StaticBearerAuthenticationHeaderValueProvider(bearerToken, logicalCluster, identityPoolId);
-                    break;
-
-                case "OAUTHBEARER":
-                    clientId = config.FirstOrDefault(prop =>
-                        prop.Key.ToLower() == SchemaRegistryConfig.PropertyNames.SchemaRegistryBearerAuthClientId).Value;
-
-                    clientSecret = config.FirstOrDefault(prop =>
-                        prop.Key.ToLower() == SchemaRegistryConfig.PropertyNames.SchemaRegistryBearerAuthClientSecret).Value;
-
-                    scope = config.FirstOrDefault(prop =>
-                        prop.Key.ToLower() == SchemaRegistryConfig.PropertyNames.SchemaRegistryBearerAuthScope).Value;
-                    
-                    tokenEndpointUrl = config.FirstOrDefault(prop =>
-                        prop.Key.ToLower() == SchemaRegistryConfig.PropertyNames.SchemaRegistryBearerAuthTokenEndpointUrl).Value;
-                    
-                    if (tokenEndpointUrl == null || clientId == null || clientSecret == null || scope == null)
-                    {
-                        throw new ArgumentException(
-                            $"Invalid bearer authentication provider configuration: Token endpoint URL, client ID, client secret, and scope must be specified");
-                    }
-                    authenticationHeaderValueProvider = new BearerAuthenticationHeaderValueProvider(
-                        new HttpClient(), clientId, clientSecret, scope, tokenEndpointUrl, logicalCluster, identityPoolId, maxRetries, retriesWaitMs, retriesMaxWaitMs);
-                    break;
-
-                case "CUSTOM":
-                    if (authenticationHeaderValueProvider == null)
-                    {
-                        throw new ArgumentException(
-                            $"Invalid authentication header value provider configuration: Custom authentication provider must be specified");
-                    }
-                    if(!(authenticationHeaderValueProvider is IAuthenticationBearerHeaderValueProvider))
-                    {
-                        throw new ArgumentException(
-                            $"Invalid authentication header value provider configuration: Custom authentication provider must implement IAuthenticationBearerHeaderValueProvider");
-                    }
-                    break;
-
-                case "":
-                    break;
-
-                default:
-                    throw new ArgumentException(
-                        $"Invalid value '{bearerAuthSource}' specified for property '{SchemaRegistryConfig.PropertyNames.SchemaRegistryBearerAuthCredentialsSource}'");
-            }
+            authenticationHeaderValueProvider = RestService.AuthenticationHeaderValueProvider(
+                config, authenticationHeaderValueProvider, maxRetries, retriesWaitMs, retriesMaxWaitMs);
 
             foreach (var property in config)
             {
